backend/requirements.txt
alembic==1.16.4
argon2-cffi==23.1.0
beautifulsoup4==4.13.4
email-validator==2.2.0
fastapi==0.116.1
playwright==1.52.0
PyJWT==2.9.0
pydantic-settings==2.10.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.42
uvicorn[standard]==0.35.0

pytest==8.4.1
